+++ src/compiler/LexicalScope.h
@@ -14,7 +14,7 @@ public:
         : outer(outer), arguments(std::move(arguments)),
           locals(std::move(locals)) {}
 
-    inline size_t GetNumberOfArguments() const { return arguments.size(); }
+    inline uint8_t GetNumberOfArguments() const { return arguments.size(); }
 
     inline size_t GetNumberOfLocals() const { return locals.size(); }
