+++ src/vmobjects/VMPrimitive.h
@@ -77,7 +77,7 @@ public:
 
     std::string AsDebugString() const override;
 
-    inline size_t GetNumberOfArguments() const final {
+    inline uint8_t GetNumberOfArguments() const final {
         return Signature::GetNumberOfArguments(load_ptr(signature));
     }
